refactor(DownloadFilesButton): Remove useless code
